chore: fix retrieve update role listing endpoint
